style(frontend): assets container reworked
