+++ packages/frontend/components/Borrow/Borrow.tsx
@@ -377,7 +377,6 @@ function Borrow({ isEditing, basePosition }: BorrowProps) {
         onClose={() => setIsConfirmationModalShown(false)}
         basePosition={basePosition}
         transactionMeta={transactionMeta}
-        isEditing={isEditing}
         actionType={actionType}
         action={() => {
           setIsConfirmationModalShown(false);
@@ -389,7 +388,3 @@ function Borrow({ isEditing, basePosition }: BorrowProps) {
 }
 
 export default Borrow;
-
-Borrow.defaultProps = {
-  position: false,
-};
