+++ packages/frontend/components/Shared/ConfirmTransactionModal.tsx
@@ -1,7 +1,6 @@
 import CloseIcon from '@mui/icons-material/Close';
 import { Box, Button, Dialog, Paper, Stack, Typography } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
-import { RoutingStep } from '@x-fuji/sdk';
 import Image from 'next/image';
 import React, { useMemo } from 'react';
 
@@ -19,7 +18,6 @@ import {
 } from '../../helpers/transactions';
 import { formatValue } from '../../helpers/values';
 import { useBorrow } from '../../store/borrow.store';
-import AssetBox from './ConfirmationTransaction/AssetBox';
 import InfoRow from './ConfirmationTransaction/InfoRow';
 import RouteBox from './ConfirmationTransaction/RouteBox';
 import WarningInfo from './WarningInfo';
@@ -28,23 +26,14 @@ type ConfirmTransactionModalProps = {
   basePosition: BasePosition;
   transactionMeta: TransactionMeta;
   open: boolean;
-  isEditing: boolean;
   actionType: ActionType;
   onClose: () => void;
   action: () => void;
 };
 
-const routingSteps = [
-  RoutingStep.DEPOSIT,
-  RoutingStep.WITHDRAW,
-  RoutingStep.BORROW,
-  RoutingStep.PAYBACK,
-];
-
 export function ConfirmTransactionModal({
   basePosition,
   transactionMeta,
-  isEditing,
   open,
   onClose,
   action,
@@ -107,14 +96,14 @@ export function ConfirmTransactionModal({
         },
         '& .MuiDialog-paper': {
           maxWidth: '40rem',
-          width: { xs: '80%', sm: '30rem' },
+          width: { xs: '80%', sm: '32rem' },
         },
       }}
     >
       <Paper
         variant="outlined"
         sx={{
-          width: { xs: 'auto', sm: '30rem' },
+          width: { xs: 'auto', sm: '32rem' },
           p: { xs: '1rem', sm: '1.5rem' },
           textAlign: 'center',
         }}
@@ -130,14 +119,6 @@ export function ConfirmTransactionModal({
         <Typography mb="1rem" variant="h5" color={palette.text.primary}>
           Confirm Transaction
         </Typography>
-
-        <Stack>
-          {steps
-            .filter((step) => routingSteps.includes(step.step))
-            .map((step) => (
-              <AssetBox key={step.step} isEditing={isEditing} step={step} />
-            ))}
-        </Stack>
 
         {steps && steps.length > 0 && (
           <RouteBox steps={steps} isCrossChain={isCrossChain} />
